+++ lang/syn/src/lib.rs
@@ -896,6 +896,7 @@ pub enum ConstraintToken {
     ExtensionTokenHookProgramId(Context<ConstraintExtensionTokenHookProgramId>),
     ExtensionPermanentDelegate(Context<ConstraintExtensionPermanentDelegate>),
     ExtensionPausableAuthority(Context<ConstraintExtensionAuthority>),
+    ExtensionImmutableOwner(Context<ConstraintExtensionImmutableOwner>),
 }
 
 impl Parse for ConstraintToken {
@@ -1133,6 +1134,11 @@ pub struct ConstraintExtensionPermanentDelegate {
 }
 
 #[derive(Debug, Clone)]
+pub struct ConstraintExtensionImmutableOwner {
+    pub immutable_owner: bool,
+}
+
+#[derive(Debug, Clone)]
 #[allow(clippy::large_enum_variant)]
 pub enum InitKind {
     Program {
@@ -1147,6 +1153,8 @@ pub enum InitKind {
         owner: Expr,
         mint: Expr,
         token_program: Option<Expr>,
+        // extensions
+        immutable_owner: Option<bool>,
     },
     AssociatedToken {
         owner: Expr,
@@ -1270,6 +1278,7 @@ pub struct ConstraintTokenAccountGroup {
     pub mint: Option<Expr>,
     pub authority: Option<Expr>,
     pub token_program: Option<Expr>,
+    pub immutable_owner: Option<bool>,
 }
 
 #[derive(Debug, Clone)]